% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_setupLocalJBrowseR.R
\name{download_genome_reference}
\alias{download_genome_reference}
\title{Download reference genome for local JBrowseR server}
\usage{
download_genome_reference(browser_server_path, genome_version)
}
\description{
Download reference genome for local JBrowseR server
}
